+++ jib-cli/src/test/java/com/google/cloud/tools/jib/cli/JibCliTest.java
@@ -127,7 +127,8 @@ public class JibCliTest {
       throws InvalidImageReferenceException, IOException, DigestException {
     String imageId = "sha256:61bb3ec31a47cb730eb58a38bbfa813761a51dca69d10e39c24c3d00a7b2c7a9";
     String digest = "sha256:3f1be7e19129edb202c071a659a4db35280ab2bb1a16f223bfd5d1948657b6fc";
-    when(mockJibContainer.getTargetImage()).thenReturn(ImageReference.parse("adoptopenjdk:8-jre"));
+    when(mockJibContainer.getTargetImage())
+        .thenReturn(ImageReference.parse("eclipse-temurin:8-jre"));
     when(mockJibContainer.getImageId()).thenReturn(DescriptorDigest.fromDigest(imageId));
     when(mockJibContainer.getDigest()).thenReturn(DescriptorDigest.fromDigest(digest));
     when(mockJibContainer.getTags()).thenReturn(ImmutableSet.of("latest", "tag-2"));
@@ -138,7 +139,7 @@ public class JibCliTest {
     String outputJson = new String(Files.readAllBytes(outputPath), StandardCharsets.UTF_8);
     ImageMetadataOutput metadataOutput =
         JsonTemplateMapper.readJson(outputJson, ImageMetadataOutput.class);
-    assertThat(metadataOutput.getImage()).isEqualTo("adoptopenjdk:8-jre");
+    assertThat(metadataOutput.getImage()).isEqualTo("eclipse-temurin:8-jre");
     assertThat(metadataOutput.getImageId()).isEqualTo(imageId);
     assertThat(metadataOutput.getImageDigest()).isEqualTo(digest);
     assertThat(metadataOutput.getTags()).containsExactly("latest", "tag-2");