+++ jib-cli/src/test/java/com/google/cloud/tools/jib/cli/jar/JarFilesTest.java
@@ -40,35 +40,42 @@ import java.nio.file.Paths;
 import java.time.Instant;
 import java.util.Arrays;
 import java.util.Optional;
+import junitparams.JUnitParamsRunner;
+import junitparams.Parameters;
+import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.MockitoJUnit;
+import org.mockito.junit.MockitoRule;
 
 /** Tests for {@link JarFiles}. */
-@RunWith(MockitoJUnitRunner.class)
+@RunWith(JUnitParamsRunner.class)
 public class JarFilesTest {
 
+  @Rule public final MockitoRule mockitoRule = MockitoJUnit.rule().silent();
+
   @Mock private StandardExplodedProcessor mockStandardExplodedProcessor;
-
   @Mock private StandardPackagedProcessor mockStandardPackagedProcessor;
-
   @Mock private SpringBootExplodedProcessor mockSpringBootExplodedProcessor;
-
   @Mock private SpringBootPackagedProcessor mockSpringBootPackagedProcessor;
-
   @Mock private Jar mockJarCommand;
-
   @Mock private CommonCliOptions mockCommonCliOptions;
-
   @Mock private CommonContainerConfigCliOptions mockCommonContainerConfigCliOptions;
-
   @Mock private ConsoleLogger mockLogger;
 
   @Test
-  public void testToJibContainer_defaultBaseImage_java8()
+  @Parameters(
+      value = {
+        "8, eclipse-temurin:8-jre",
+        "9, eclipse-temurin:11-jre",
+        "11, eclipse-temurin:11-jre",
+        "13, azul/zulu-openjdk:17-jre",
+        "17, azul/zulu-openjdk:17-jre",
+      })
+  public void testToJibContainer_defaultBaseImage(int javaVersion, String expectedBaseImage)
       throws IOException, InvalidImageReferenceException {
-    when(mockStandardExplodedProcessor.getJavaVersion()).thenReturn(8);
+    when(mockStandardExplodedProcessor.getJavaVersion()).thenReturn(javaVersion);
     JibContainerBuilder containerBuilder =
         JarFiles.toJibContainerBuilder(
             mockStandardExplodedProcessor,
@@ -78,23 +85,7 @@ public class JarFilesTest {
             mockLogger);
     ContainerBuildPlan buildPlan = containerBuilder.toContainerBuildPlan();
 
-    assertThat(buildPlan.getBaseImage()).isEqualTo("adoptopenjdk:8-jre");
-  }
-
-  @Test
-  public void testToJibContainer_defaultBaseImage_java9()
-      throws IOException, InvalidImageReferenceException {
-    when(mockStandardExplodedProcessor.getJavaVersion()).thenReturn(9);
-    JibContainerBuilder containerBuilder =
-        JarFiles.toJibContainerBuilder(
-            mockStandardExplodedProcessor,
-            mockJarCommand,
-            mockCommonCliOptions,
-            mockCommonContainerConfigCliOptions,
-            mockLogger);
-    ContainerBuildPlan buildPlan = containerBuilder.toContainerBuildPlan();
-
-    assertThat(buildPlan.getBaseImage()).isEqualTo("adoptopenjdk:11-jre");
+    assertThat(buildPlan.getBaseImage()).isEqualTo(expectedBaseImage);
   }
 
   @Test
@@ -123,7 +114,7 @@ public class JarFilesTest {
             mockLogger);
     ContainerBuildPlan buildPlan = containerBuilder.toContainerBuildPlan();
 
-    assertThat(buildPlan.getBaseImage()).isEqualTo("adoptopenjdk:8-jre");
+    assertThat(buildPlan.getBaseImage()).isEqualTo("eclipse-temurin:8-jre");
     assertThat(buildPlan.getPlatforms()).isEqualTo(ImmutableSet.of(new Platform("amd64", "linux")));
     assertThat(buildPlan.getCreationTime()).isEqualTo(Instant.EPOCH);
     assertThat(buildPlan.getFormat()).isEqualTo(ImageFormat.Docker);
@@ -172,7 +163,7 @@ public class JarFilesTest {
             mockLogger);
     ContainerBuildPlan buildPlan = containerBuilder.toContainerBuildPlan();
 
-    assertThat(buildPlan.getBaseImage()).isEqualTo("adoptopenjdk:8-jre");
+    assertThat(buildPlan.getBaseImage()).isEqualTo("eclipse-temurin:8-jre");
     assertThat(buildPlan.getPlatforms()).isEqualTo(ImmutableSet.of(new Platform("amd64", "linux")));
     assertThat(buildPlan.getCreationTime()).isEqualTo(Instant.EPOCH);
     assertThat(buildPlan.getFormat()).isEqualTo(ImageFormat.Docker);
@@ -223,7 +214,7 @@ public class JarFilesTest {
             mockLogger);
     ContainerBuildPlan buildPlan = containerBuilder.toContainerBuildPlan();
 
-    assertThat(buildPlan.getBaseImage()).isEqualTo("adoptopenjdk:8-jre");
+    assertThat(buildPlan.getBaseImage()).isEqualTo("eclipse-temurin:8-jre");
     assertThat(buildPlan.getPlatforms()).isEqualTo(ImmutableSet.of(new Platform("amd64", "linux")));
     assertThat(buildPlan.getCreationTime()).isEqualTo(Instant.EPOCH);
     assertThat(buildPlan.getFormat()).isEqualTo(ImageFormat.Docker);
@@ -272,7 +263,7 @@ public class JarFilesTest {
             mockLogger);
     ContainerBuildPlan buildPlan = containerBuilder.toContainerBuildPlan();
 
-    assertThat(buildPlan.getBaseImage()).isEqualTo("adoptopenjdk:8-jre");
+    assertThat(buildPlan.getBaseImage()).isEqualTo("eclipse-temurin:8-jre");
     assertThat(buildPlan.getPlatforms()).isEqualTo(ImmutableSet.of(new Platform("amd64", "linux")));
     assertThat(buildPlan.getCreationTime()).isEqualTo(Instant.EPOCH);
     assertThat(buildPlan.getFormat()).isEqualTo(ImageFormat.Docker);